refactor: Separate datatable pagination from data table file
